Refactor code (so that it actually runs)...
- Invoke does not have access to the local virtual environment
- Some functions need to be split out from settings.py
