+++ requirements.txt
@@ -33,7 +33,6 @@ flake8==3.8.3                   # PEP checking
 gunicorn>=20.1.0                # Gunicorn web server
 importlib_metadata              # Backport for importlib.metadata
 inventree                       # Install the latest version of the InvenTree API python library
-invoke>=1.6.0                   # Invoke management tool (must be installed in venv)
 markdown==3.3.4                 # Force particular version of markdown
 pep8-naming==0.11.1             # PEP naming convention extension
 pillow==8.3.2                   # Image manipulation
